move implementation of read-only-issue.body txt, and tests to bitbucket platform
