+++ lib/modules/platform/bitbucket/__fixtures__/issue-body.txt
