--- lib/modules/platform/utils/read-only-issue-body.spec.ts
@@ -1,22 +1,0 @@
-import { Fixtures } from '../../../../test/fixtures';
-import { readOnlyIssueBody } from './read-only-issue-body';
-
-const issueBody = Fixtures.get('issue-body.txt');
-
-describe('modules/platform/utils/read-only-issue-body', () => {
-  describe('.readOnlyIssueBody', () => {
-    it('removes all checkbox formatting', () => {
-      expect(readOnlyIssueBody(issueBody)).toEqual(
-        expect.not.stringContaining('[ ] <!--')
-      );
-    });
-
-    it('removes all checkbox-related instructions', () => {
-      expect(readOnlyIssueBody(issueBody)).toEqual(
-        expect.not.stringMatching(
-          /click (?:(?:on |)a|their|this) checkbox|check the box below/gi
-        )
-      );
-    });
-  });
-});
